refactor: remove stop-on-first param, and tests cases cleanup
