+++ test/returns_interface_type_test.go
@@ -8,13 +8,8 @@ import (
 )
 
 func TestReturnsInterfaceTypeDefault(t *testing.T) {
-	testRule(t, "returns_interface_type_default", &rule.ReturnsInterfaceTypeRule{}, &lint.RuleConfig{})
-}
-
-func TestReturnsInterfaceTypeSkip(t *testing.T) {
-	testRule(t, "returns_interface_type_skip", &rule.ReturnsInterfaceTypeRule{}, &lint.RuleConfig{
+	testRule(t, "returns_interface_type_default", &rule.ReturnsInterfaceTypeRule{}, &lint.RuleConfig{
 		Arguments: lint.Arguments{map[string]any{
-			"stopOnFirst": true,
 			"ignoredNames": []any{
 				"fixtures.DummyResults",
 			},
