+++ testdata/returns_interface_type_default.go
@@ -34,8 +34,35 @@ func (a A) B() DummyReader { // MATCH /fixtures.A.B returns interface type fixtu
 	return nil
 }
 
-type DummyReader interface{}
+func firstSkip() (DummyWriter, error) { // MATCH /firstSkip returns interface type fixtures.DummyWriter/
+	return nil, nil
+}
+
+func skipDummyResults() DummyResults {
+	return nil
+}
+
+func spotMiddle() (any, DummyWriter, error) { // MATCH /spotMiddle returns interface type fixtures.DummyWriter/
+	return nil, nil, nil
+}
+
+func spotLast() (any, DummyWriter) { // MATCH /spotLast returns interface type fixtures.DummyWriter/
+	return nil, nil
+}
+
+func skipAll() (DummyResults, any, interface{}, error) {
+	return nil, nil, nil, nil
+}
 
 type Reader interface {
 	Read([]byte) (int, error)
+}
+type DummyReader interface{}
+type DummyResults interface{}
+type DummyWriter interface{}
+
+type Dummy struct{}
+
+func (d Dummy) returnSkippedInterfaceType() DummyResults {
+	return nil
 }
